Refactor link-pred example and remove white spaces
